move new files after rebase
