+++ apis/sgv2-docsapi/src/main/java/io/stargate/sgv2/docsapi/api/v2/namespaces/collections/model/dto/JsonSchemaDto.java
