+++ apis/sgv2-docsapi/src/main/java/io/stargate/sgv2/docsapi/service/function/BuiltInFunction.java
